requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.10.0
numpy==1.23.5
nltk==3.8.1
